Move agent::tests
